Sunburst chart (#46)
I started off with Plotly.js, then got frustrated with its API and
limitations, then tried Echarts which was better but still didn't
provide the options and simplicity I wanted, so I eventually just wrote
it myself. I feel this looks better and will be more customizable, and
is actually more simple/predictable for almost the same amount of code.
One perhaps key thing the libraries do that my custom version doesn't
(yet) is allow you to click on a slice and "expand" it. If we needed
that it'd then become more code than using the libraries, but I'm not
sure how useful that is.

- pull in misc small component updates from geneplexus
- in network component:
- use reactive color map for node/edge colors like everywhere else in
code base, instead of "colorMix" util function
  - fix bug where colors wouldn't update when dark mode changes
  - truncate edge labels based on distance between nodes
- add custom sunburst chart component
- add fake data for testbed test of sunburst
- refactor color palette, re-choose colors
- move around some util funcs

I made some changes to the color util funcs to make the sunburst chart
look good, which led to me needing to change the network component as
well.