Interpro redo (#53)
Compare the Netlify preview here to the `main` one at
https://molevolvr.netlify.app/testbed, where I re-enabled the
Nightingale IPR viz so you could compare. Note that it [seems to slow
down the
page](https://github.com/JRaviLab/molevolvr2.0/issues/7#issuecomment-2877223451),
which is why I had commented it out originally while developing other
components.

- remove bespoke "install playwright" action, due to me observing it
causing problems [in this
repo](https://github.com/nih-cfde/icc-eval-core), and not being able to
get it to reliably save any time
- get rid of eslint config compatibility lines, clean up
- update all package versions (all minor bumps)
- delete nightingale packages
- on CI, only run e2e tests in chromium
- css var name changes, remove vestigial styles, misc tweaks
- develop new interpro visualization component using d3
- reduce testbed msa data to reduce # of elements on page at same time.
in real life, msa wont be on screen at the same time as every other
component we have, like it is on the testbed