Upset plot, svg/charting overhaul (#55)
Unfortunately this became a big PR because everything here was
intertwined, and I wanted to (hopefully) solve all these related issues
permanently. So, the diff is very big and difficult to review, and the
individual commits won't be very clean or helpful. Instead I will mark
certain parts of the code to look at. And know that converting the old
SVGs to use the new chart wrapper component involved a lot of
copy-pasting large blocks of code. Most importantly, please re-test
every visualization component on the Netlify preview.

At a high level, here are the changes:

- implement upset plot
- make generic chart wrapper component that handles most things
automatically. auto-fits view box to content. SVG units match DOM units
1:1. see discussion below for how i landed on this particular design.
- abstract out download button with various download formats as its own
component. integrate into chart component.
- chart text can be truncated automatically by actual rendered width
(not number of chars)
- include chart titles in every chart
- include analysis id (or anything else we want) as part of every
filename download. add fake analysis id to testbed page.
- turn the legend component into fully SVG instead of DOM elements.
implement simple wrapping strategy.
- make MSA and IPR charts fully SVG, with no DOM elements
- MSA now auto-wraps by default, and accurately (instead of loose
estimates). when printing,
- make charts resizable
- add SVG group elements where appropriate for easier editing in vector
software
- consistently put settings vars and Props type at the top of each
component file for quick reference
- util func simplifications and refactors

---

If you poke through the commit history, you can see some of the many
complicated approaches I experimented with for handling sizing and
positioning in the SVGs.

I wanted to scale the SVG shapes up/down based on the available width
but keep the font size matching the document's. This was already sort of
implemented before this PR, where the viz components made repeated use
of a useSvgTransform hook. But it was verbose, so I first attempted to
extract it out to a single hook, and then a component. It required
iteratively content-fitting and font-sizing back and forth, which would
(usually) converge to a stable configuration. But it was very difficult
to find a foolproof way to prevent infinite re-renders or weird edge
cases. I also tried leveraging simply defining font size in the root of
the SVG that matched the document's, then specifying positions/sizes in
terms of `em` units, which the browser could use synchronously with no
JS. But there are some limitations to this, like not being able to use
relative units in things like `transform`s, and some weird exceptions
that caused runaway calculations like when an `em`-sized element
exceeded the available width of the container. Trying to implement
automatic text truncation made this even more complicated.

The MSA & IPR viz's also required a "fill remaining space" feature, i.e.
have a column of labels on the left of a fixed width, and use all the
remaining space on the right for the tracks/sequences. Another small
problem to solve was that `getBBox`, the method used for fitting the
view box to the SVG contents, doesn't work with `clip-path`'ed elements,
which is needed for IPR's zoom area.

There's much more that I tried or thought about and didn't commit, which
has already left my brain, but I think it's not that valuable.

What I landed on keeps things much simpler. The SVG units always match
the DOM units, no exception (hopefully). If there isn't enough room on
the page to show the SVG's content, instead of trying to shrink
dynamically, it will just overflow and show scrollbars. Not only does
this avoid needing iterative fitting, it's probably better UX wise since
graphical elements stay in the same position and size -- instead of
shrinking to an unreadable size -- and you just need to scroll over to
see it all. Also this makes the testbed page much more performant. The
"fill remaining" feature is accomplished by providing the available
width to the consuming component, and letting it use it however it
wants.

---

Regarding the MSA auto-wrapping, currently the code is basically:

```ts
setWrap(40);
print()
setWrap(oldWrap);
```

This is just a loose, hard-coded wrapping point that would generally not
overflow the page width when printing portrait mode, US letter.

This PR doesn't change our inability to look inside the print dialog, as
discussed in #50 and [this
comment](https://github.com/JRaviLab/molevolvr2.0/pull/52#issuecomment-2874220380).
But I've taken out the hard-coded `40`. I figure now with the resize
handle available on every chart, the user can pre-size the chart to
whatever they want before opening the print dialog.

I wanted to mention this because, if you look at the preview on a normal
desktop or laptop screen width and have your print dialog on defaults,
you'll usually see the MSA overflow the page, and I'm aware of it.