Table of contents and about page rework (#59)
- simplify flex component props, e.g. `direction="column"` -> `column`,
since it's used everywhere. find+replace whole app.
- change heading component to register/unregister itself in global list
of headings
- consistent use of data attribute for active state styles
- tweak table of contents styles
- instead of requerying for all headings on any document mutation
(expensive\*), render table of contents from global reactive list of
headings
- include icons in table of contents entries
- clean up and compactify about page (original impetus for making TOC
look and behave better)

\* On any modern machine though, probably not a humanly perceptible
difference. Also the expense will be limited by the fact that we should
never have as many things on one page as is currently on the testbed
page.