Feedback form (#61)
Closes #56 

- pull out repeated static text snippets -- like lab name, email, github
org, etc -- into `.env` file variables
- add type defs for env vars
- add random delay to API mock handlers to test loading spinners and
such
- move tanstack-query QueryProvider up in app tree so feedback form can
use it
- add placeholder api func for submitting feedback form
- tweak dialog component props and styles, including
on-open-state-change callback
- add feedback form component in corner of every page
    - optional name, username, email fields
- sync fields with local storage, reset on successful form submission
    - auto-include browser/OS/device/etc details
- in case of error, fallback to `mailto:` link with pre-filled subject
and body
- use new .env vars wherever appropriate, such as in footer links
- fix link component bug when navigating to root page
- tweak alert component types and styles
- use destructured `import.meta.env` wherever appropriate
- in textbox component, only show copy button on multiline inputs
- make "view corner" component just buttons. move toasts to separate
element.
- make note on about page to encourage use of feedback form instead of
directly emailing

@falquaddoomi I was initially thinking we'd just do a one-off
serverless/cloud function for this. But maybe since we'll have a
continuously running server anyway, we can just make this part of the
regular backend. But on the other hand maybe it'd be good to have the
feedback form as an isolated thing so if the backend goes down for some
reason the feedback form still works? If you want to do it as a cloud
function, I have a snippet of Node code ready to go (from another
project) that I could set up ahead of time. I'll leave it up to you.

---------

Co-authored-by: Faisal Alquaddoomi <falquaddoomi@gmail.com>